requirements.txt
python-telegram-bot==21.0.0
Flask==2.3.3
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
pytz==2023.3.post1
twilio==8.12.0
gunicorn==23.0.0
anthropic==0.22.1
openai==1.13.3